+++ contracts/ibc-reflect/tests/integration.rs
@@ -91,11 +91,13 @@ fn connect(
         res.events[0]
     );
     let id = res.messages[0].id;
+    let payload = res.messages[0].payload.clone();
 
     // fake a reply and ensure this works
     #[allow(deprecated)]
     let response = Reply {
         id,
+        payload,
         gas_used: 1234567,
         result: SubMsgResult::Ok(SubMsgResponse {
             events: fake_events(&account),
@@ -153,6 +155,7 @@ fn proper_handshake_flow() {
     // and set up a reflect account
     assert_eq!(1, res.messages.len());
     let id = res.messages[0].id;
+    let payload = res.messages[0].payload.clone();
     if let CosmosMsg::Wasm(WasmMsg::Instantiate {
         admin,
         code_id,
@@ -178,6 +181,7 @@ fn proper_handshake_flow() {
     #[allow(deprecated)]
     let response = Reply {
         id,
+        payload,
         gas_used: 1234567,
         result: SubMsgResult::Ok(SubMsgResponse {
             events: fake_events(reflect_addr.as_str()),