+++ contracts/reflect/src/contract.rs
@@ -434,6 +434,7 @@ mod tests {
         let _res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
 
         let id = 123u64;
+        let payload = Binary::from(b"my dear");
         let data = Binary::from(b"foobar");
         let events = vec![Event::new("message").add_attribute("signer", "caller-addr")];
         let gas_used = 1234567u64;
@@ -443,12 +444,13 @@ mod tests {
             data: Some(data.clone()),
             msg_responses: vec![],
         });
-        let subcall = Reply {
+        let the_reply = Reply {
             id,
+            payload,
             gas_used,
             result,
         };
-        let res = reply(deps.as_mut(), mock_env(), subcall).unwrap();
+        let res = reply(deps.as_mut(), mock_env(), the_reply).unwrap();
         assert_eq!(0, res.messages.len());
 
         // query for a non-existant id