+++ packages/std/src/results/submessages.rs
@@ -33,8 +33,25 @@ pub struct SubMsg<T = Empty> {
     /// An arbitrary ID chosen by the contract.
     /// This is typically used to match `Reply`s in the `reply` entry point to the submessage.
     pub id: u64,
+    /// Some arbirary data that the contract can set in an application specific way.
+    /// This is just passed into the `reply` entry point and is not stored to state.
+    /// Any encoding can be used. If `id` is used to identify a particular action,
+    /// the encoding can also be different for each of those actions since you can match `id`
+    /// first and then start processing the `payload`.
+    ///
+    /// The environment restricts the length of this field in order to avoid abuse. The limit
+    /// is environment specific and can change over time. The initial default is 128 KiB.
+    ///
+    /// Unset/nil/null cannot be differentiated from empty data.
+    ///
+    /// On chains running CosmWasm 1.x this field will be ignored.
+    #[serde(default)]
+    pub payload: Binary,
     pub msg: CosmosMsg<T>,
     /// Gas limit measured in [Cosmos SDK gas](https://github.com/CosmWasm/cosmwasm/blob/main/docs/GAS.md).
+    ///
+    /// Setting this to `None` means unlimited. Then the submessage execution can consume all gas of the
+    /// current execution context.
     pub gas_limit: Option<u64>,
     pub reply_on: ReplyOn,
 }
@@ -43,32 +60,50 @@ pub struct SubMsg<T = Empty> {
 pub const UNUSED_MSG_ID: u64 = 0;
 
 impl<T> SubMsg<T> {
-    /// new creates a "fire and forget" message with the pre-0.14 semantics
+    /// Creates a "fire and forget" message with the pre-0.14 semantics.
+    /// Since this is just an alias for [`SubMsg::reply_never`] it is somewhat recommended
+    /// to use the latter in order to make the behaviour more explicit in the caller code.
+    /// But that's up to you for now.
+    ///
+    /// By default, the submessage's gas limit will be unlimited. Use [`SubMsg::with_gas_limit`] to change it.
+    /// Setting `payload` is not advised as this will never be used.
     pub fn new(msg: impl Into<CosmosMsg<T>>) -> Self {
         Self::reply_never(msg)
     }
 
-    /// create a `SubMsg` that will provide a `reply` with the given id if the message returns `Ok`
+    /// Creates a `SubMsg` that will provide a `reply` with the given `id` if the message returns `Ok`.
+    ///
+    /// By default, the submessage's `payload` will be empty and the gas limit will be unlimited. Use
+    /// [`SubMsg::with_payload`] and [`SubMsg::with_gas_limit`] to change those.
     pub fn reply_on_success(msg: impl Into<CosmosMsg<T>>, id: u64) -> Self {
         Self::reply_on(msg.into(), id, ReplyOn::Success)
     }
 
-    /// create a `SubMsg` that will provide a `reply` with the given id if the message returns `Err`
+    /// Creates a `SubMsg` that will provide a `reply` with the given `id` if the message returns `Err`.
+    ///
+    /// By default, the submessage's `payload` will be empty and the gas limit will be unlimited. Use
+    /// [`SubMsg::with_payload`] and [`SubMsg::with_gas_limit`] to change those.
     pub fn reply_on_error(msg: impl Into<CosmosMsg<T>>, id: u64) -> Self {
         Self::reply_on(msg.into(), id, ReplyOn::Error)
     }
 
-    /// create a `SubMsg` that will always provide a `reply` with the given id
+    /// Create a `SubMsg` that will always provide a `reply` with the given `id`.
+    ///
+    /// By default, the submessage's `payload` will be empty and the gas limit will be unlimited. Use
+    /// [`SubMsg::with_payload`] and [`SubMsg::with_gas_limit`] to change those.
     pub fn reply_always(msg: impl Into<CosmosMsg<T>>, id: u64) -> Self {
         Self::reply_on(msg.into(), id, ReplyOn::Always)
     }
 
-    /// create a `SubMsg` that will never `reply`. This is equivalent to standard message semantics.
+    /// Create a `SubMsg` that will never `reply`. This is equivalent to standard message semantics.
+    ///
+    /// By default, the submessage's gas limit will be unlimited. Use [`SubMsg::with_gas_limit`] to change it.
+    /// Setting `payload` is not advised as this will never be used.
     pub fn reply_never(msg: impl Into<CosmosMsg<T>>) -> Self {
         Self::reply_on(msg.into(), UNUSED_MSG_ID, ReplyOn::Never)
     }
 
-    /// Add a gas limit to the message.
+    /// Add a gas limit to the submessage.
     /// This gas limit measured in [Cosmos SDK gas](https://github.com/CosmWasm/cosmwasm/blob/main/docs/GAS.md).
     ///
     /// ## Examples
@@ -86,9 +121,28 @@ impl<T> SubMsg<T> {
         self
     }
 
+    /// Add a payload to the submessage.
+    ///
+    /// ## Examples
+    ///
+    /// ```
+    /// # use cosmwasm_std::{coins, BankMsg, Binary, ReplyOn, SubMsg};
+    /// # let msg = BankMsg::Send { to_address: String::from("you"), amount: coins(1015, "earth") };
+    /// let sub_msg: SubMsg = SubMsg::reply_always(msg, 1234)
+    ///     .with_payload(vec![1, 2, 3, 4]);
+    /// assert_eq!(sub_msg.id, 1234);
+    /// assert_eq!(sub_msg.payload, Binary::new(vec![1, 2, 3, 4]));
+    /// assert_eq!(sub_msg.reply_on, ReplyOn::Always);
+    /// ```
+    pub fn with_payload(mut self, payload: impl Into<Binary>) -> Self {
+        self.payload = payload.into();
+        self
+    }
+
     fn reply_on(msg: CosmosMsg<T>, id: u64, reply_on: ReplyOn) -> Self {
         SubMsg {
             id,
+            payload: Default::default(),
             msg,
             reply_on,
             gas_limit: None,
@@ -103,6 +157,14 @@ pub struct Reply {
     /// The ID that the contract set when emitting the `SubMsg`.
     /// Use this to identify which submessage triggered the `reply`.
     pub id: u64,
+    /// Some arbirary data that the contract set when emitting the `SubMsg`.
+    /// This is just passed into the `reply` entry point and is not stored to state.
+    ///
+    /// Unset/nil/null cannot be differentiated from empty data.
+    ///
+    /// On chains running CosmWasm 1.x this field is never filled.
+    #[serde(default)]
+    pub payload: Binary,
     /// The amount of gas used by the submessage,
     /// measured in [Cosmos SDK gas](https://github.com/CosmWasm/cosmwasm/blob/main/docs/GAS.md).
     pub gas_used: u64,
@@ -221,7 +283,71 @@ pub struct MsgResponse {
 #[allow(deprecated)]
 mod tests {
     use super::*;
-    use crate::{from_json, to_json_vec, StdError, StdResult};
+    use crate::{coins, from_json, to_json_vec, Attribute, BankMsg, StdError, StdResult};
+
+    #[test]
+    fn sub_msg_new_works() {
+        let msg = BankMsg::Send {
+            to_address: String::from("you"),
+            amount: coins(1015, "earth"),
+        };
+        let sub_msg: SubMsg = SubMsg::new(msg.clone());
+        // id and payload don't matter since there is no reply
+        assert_eq!(sub_msg.reply_on, ReplyOn::Never);
+        assert_eq!(sub_msg.gas_limit, None);
+        assert_eq!(sub_msg.msg, CosmosMsg::from(msg));
+    }
+
+    #[test]
+    fn sub_msg_reply_never_works() {
+        let msg = BankMsg::Send {
+            to_address: String::from("you"),
+            amount: coins(1015, "earth"),
+        };
+        let sub_msg: SubMsg = SubMsg::reply_never(msg.clone());
+        // id and payload don't matter since there is no reply
+        assert_eq!(sub_msg.reply_on, ReplyOn::Never);
+        assert_eq!(sub_msg.gas_limit, None);
+        assert_eq!(sub_msg.msg, CosmosMsg::from(msg));
+    }
+
+    #[test]
+    fn sub_msg_reply_always_works() {
+        let msg = BankMsg::Send {
+            to_address: String::from("you"),
+            amount: coins(1015, "earth"),
+        };
+        let sub_msg: SubMsg = SubMsg::reply_always(msg.clone(), 54);
+        assert_eq!(sub_msg.id, 54);
+        assert_eq!(sub_msg.payload, Binary::default());
+        assert_eq!(sub_msg.reply_on, ReplyOn::Always);
+        assert_eq!(sub_msg.gas_limit, None);
+        assert_eq!(sub_msg.msg, CosmosMsg::from(msg));
+    }
+
+    #[test]
+    fn sub_msg_with_gas_limit_works() {
+        let msg = BankMsg::Send {
+            to_address: String::from("you"),
+            amount: coins(1015, "earth"),
+        };
+        let sub_msg: SubMsg = SubMsg::reply_never(msg);
+        assert_eq!(sub_msg.gas_limit, None);
+        let sub_msg = sub_msg.with_gas_limit(20);
+        assert_eq!(sub_msg.gas_limit, Some(20));
+    }
+
+    #[test]
+    fn sub_msg_with_payload_works() {
+        let msg = BankMsg::Send {
+            to_address: String::from("you"),
+            amount: coins(1015, "earth"),
+        };
+        let sub_msg: SubMsg = SubMsg::reply_never(msg);
+        assert_eq!(sub_msg.payload, Binary::default());
+        let sub_msg = sub_msg.with_payload(vec![0xAA, 3, 5, 1, 2]);
+        assert_eq!(sub_msg.payload, Binary::new(vec![0xAA, 3, 5, 1, 2]));
+    }
 
     #[test]
     fn sub_msg_result_serialization_works() {
@@ -417,5 +543,52 @@ mod tests {
         let original = SubMsgResult::Err("went wrong".to_string());
         let converted: Result<SubMsgResponse, String> = original.into();
         assert_eq!(converted, Err("went wrong".to_string()));
+    }
+
+    #[test]
+    fn reply_deserialization_works() {
+        // 1.x reply without payload (from https://github.com/CosmWasm/cosmwasm/issues/1909)
+        let reply: Reply = from_json(r#"{"gas_used":4312324,"id":75,"result":{"ok":{"events":[{"type":"hi","attributes":[{"key":"si","value":"claro"}]}],"data":"PwCqXKs="}}}"#).unwrap();
+        assert_eq!(
+            reply,
+            Reply {
+                id: 75,
+                payload: Binary::default(),
+                gas_used: 4312324,
+                result: SubMsgResult::Ok(SubMsgResponse {
+                    data: Some(Binary::from_base64("PwCqXKs=").unwrap()),
+                    events: vec![Event {
+                        ty: "hi".to_string(),
+                        attributes: vec![Attribute {
+                            key: "si".to_string(),
+                            value: "claro".to_string(),
+                        }]
+                    }],
+                    msg_responses: vec![],
+                })
+            }
+        );
+
+        // with payload (manually added to the above test)
+        let reply: Reply = from_json(r#"{"gas_used":4312324,"id":75,"payload":"3NxjC5U=","result":{"ok":{"events":[{"type":"hi","attributes":[{"key":"si","value":"claro"}]}],"data":"PwCqXKs="}}}"#).unwrap();
+        assert_eq!(
+            reply,
+            Reply {
+                id: 75,
+                payload: Binary::from_base64("3NxjC5U=").unwrap(),
+                gas_used: 4312324,
+                result: SubMsgResult::Ok(SubMsgResponse {
+                    data: Some(Binary::from_base64("PwCqXKs=").unwrap()),
+                    events: vec![Event {
+                        ty: "hi".to_string(),
+                        attributes: vec![Attribute {
+                            key: "si".to_string(),
+                            value: "claro".to_string(),
+                        }]
+                    }],
+                    msg_responses: vec![],
+                })
+            }
+        );
     }
 }